+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added support for custom transit handlers on the client side. Server side is coming in a release soon.
 - Added support for turning on/off Om path optimization
 - Fix for latest cljs support (PR 47)
+- DEPRECATED: load-data will soon no longer supports the :ident parameter. Use load instead.
 
 0.5.7
 -----